now builds x86_64 and i386, ppc is commented out as it is deprecated, use $(CC) instead of gcc white space, fix format warning

+++ authexec.c
@@ -1,65 +1,65 @@
-#include <Security/Authorization.h> 
-#include <Security/AuthorizationTags.h> 
+#include <Security/Authorization.h>
+#include <Security/AuthorizationTags.h>
 
-int read (long,StringPtr,int); 
-int write (long,StringPtr,int); 
+int read (long,StringPtr,int);
+int write (long,StringPtr,int);
 
-int main(int argc, char* argv[]) { 
-    OSStatus myStatus; 
-    AuthorizationRef myAuthorizationRef; 
+int main(int argc, char* argv[]) {
+    OSStatus myStatus;
+    AuthorizationRef myAuthorizationRef;
 
     myStatus = AuthorizationCreate(
         NULL,
         kAuthorizationEmptyEnvironment,
         kAuthorizationFlagDefaults,
-        &myAuthorizationRef); 
+        &myAuthorizationRef);
 
-    if (myStatus != errAuthorizationSuccess) 
-        return myStatus; 
+    if (myStatus != errAuthorizationSuccess)
+        return myStatus;
 
-    do 
-    { 
-        { 
-          AuthorizationItem myItems = {kAuthorizationRightExecute, 0, NULL, 0}; 
-          AuthorizationRights myRights = {1, &myItems}; 
+    do
+    {
+        {
+          AuthorizationItem myItems = {kAuthorizationRightExecute, 0, NULL, 0};
+          AuthorizationRights myRights = {1, &myItems};
           AuthorizationFlags myFlags =
-              kAuthorizationFlagDefaults | 
-              kAuthorizationFlagInteractionAllowed | 
-              kAuthorizationFlagPreAuthorize | 
+              kAuthorizationFlagDefaults |
+              kAuthorizationFlagInteractionAllowed |
+              kAuthorizationFlagPreAuthorize |
               kAuthorizationFlagExtendRights;
 
           myStatus = AuthorizationCopyRights(
-              myAuthorizationRef, &myRights, NULL, myFlags, NULL ); 
+              myAuthorizationRef, &myRights, NULL, myFlags, NULL );
         }
 
         if (myStatus != errAuthorizationSuccess)
             break;
- 
-        { 
-            FILE *myCommunicationsPipe = NULL; 
-            unsigned char myReadBuffer[128]; 
+
+        {
+            FILE *myCommunicationsPipe = NULL;
+            unsigned char myReadBuffer[128];
 
             myStatus = AuthorizationExecuteWithPrivileges(
                 myAuthorizationRef,
                 argv[1],
                 kAuthorizationFlagDefaults,
                 &argv[2],
-                &myCommunicationsPipe); 
+                &myCommunicationsPipe);
 
-            if (myStatus == errAuthorizationSuccess) 
-                for(;;) 
-                { 
-                    int bytesRead = read(fileno(myCommunicationsPipe), myReadBuffer, sizeof(myReadBuffer)); 
-                    if (bytesRead < 1) break; 
-                    write(fileno(stdout), myReadBuffer, bytesRead); 
-                } 
-        } 
-    } while (0); 
-    
-    AuthorizationFree(myAuthorizationRef, kAuthorizationFlagDefaults); 
+            if (myStatus == errAuthorizationSuccess)
+                for(;;)
+                {
+                    int bytesRead = read(fileno(myCommunicationsPipe), myReadBuffer, sizeof(myReadBuffer));
+                    if (bytesRead < 1) break;
+                    write(fileno(stdout), myReadBuffer, bytesRead);
+                }
+        }
+    } while (0);
+
+    AuthorizationFree(myAuthorizationRef, kAuthorizationFlagDefaults);
 
     if (myStatus)
-        printf("Status: %ld\n", myStatus); 
+        printf("Status: %ld\n", (long) myStatus);
 
-    return myStatus; 
-} 
+    return myStatus;
+}
